fix caching errors and add flamethrower mark 1
add all of the used files to the cache, this allows the rocket to
actually fire the model, since the model is cached.  Also add the
pyros flamethrower.  Currently it fires many dragonslaves, each one
does damage, but not DOT, also the pyro is not silenced or disarmed
for the duration.